requirements.txt
streamlit
chromadb
sentence-transformers
transformers
pymupdf
python-dotenv
